+++ device/esp_tinyusb/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - esp_tinyusb: Added ESP32H4 support
 - esp_tinyusb: Fixed an assertion failure on the GetOtherSpeedDescriptor() request for ESP32P4 when the OTG1.1 port is used
+- MSC: Added dynamic member and storage operation multitask protection
 
 ## 2.0.0